requirements.txt
apache-airflow-providers-postgres==5.8.0
psycopg2-binary==2.9.9
pandas==2.0.3
python-dotenv==1.0.0
streamlit==1.31.0